pyproject.toml
[tool.poetry]
name = "market-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.3"
celery = "^5.4.0"
rabbitmq = "^0.2.0"
dynaconf = "^3.2.6"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
sqlalchemy = "^2.0.35"
jinja2 = "^3.1.4"
sqlalchemy-utils = "^0.41.2"
pydantic-extra-types = "^2.9.0"
phonenumbers = "^8.13.47"
pydantic = {extras = ["email"], version = "^2.9.2"}
passlib = "^1.7.4"
asyncpg = "^0.30.0"
sqlmodel = "^0.0.22"
greenlet = "^3.1.1"
fastapi-mail = "^1.4.1"
python-multipart = "^0.0.16"
pyjwt = "^2.9.0"
flake8 = "^7.1.1"
black = "^24.10.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
flower = "^2.0.1"
flake8 = "^7.1.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.pytest.ini_options]
#pythonpath = [".", "src",]
#asyncio_mode = "auto"